007 UI polish

+++ src/ToledoMessage.Shared/DTOs/AddParticipantRequest.cs
@@ -1,3 +1,3 @@
 namespace ToledoMessage.Shared.DTOs;
 
-public sealed record AddParticipantRequest(decimal UserId);
+public sealed record AddParticipantRequest(long UserId);
