+++ src/ToledoMessage.Shared/DTOs/CreateGroupConversationRequest.cs
@@ -1,3 +1,3 @@
 namespace ToledoMessage.Shared.DTOs;
 
-public sealed record CreateGroupConversationRequest(string GroupName, List<decimal> ParticipantUserIds);
+public sealed record CreateGroupConversationRequest(string GroupName, List<long> ParticipantUserIds);
