+++ src/ToledoMessage.Shared/DTOs/DeviceRegistrationResult.cs
@@ -1,3 +1,3 @@
 namespace ToledoMessage.Shared.DTOs;
 
-public sealed record DeviceRegistrationResult(decimal DeviceId);
+public sealed record DeviceRegistrationResult(long DeviceId);
