+++ src/ToledoMessage/Models/EncryptedMessage.cs
@@ -6,10 +6,10 @@ public class EncryptedMessage
 {
     // ReSharper disable  NullableWarningSuppressionIsUsed
 
-    public decimal Id { get; set; }
-    public decimal ConversationId { get; set; }
-    public decimal SenderDeviceId { get; set; }
-    public decimal RecipientDeviceId { get; set; }
+    public long Id { get; set; }
+    public long ConversationId { get; set; }
+    public long SenderDeviceId { get; set; }
+    public long RecipientDeviceId { get; set; }
     public byte[] Ciphertext { get; set; } = [];
     public MessageType MessageType { get; set; }
     public ContentType ContentType { get; set; }
@@ -19,7 +19,7 @@ public class EncryptedMessage
     public DateTimeOffset ServerTimestamp { get; set; }
     public bool IsDelivered { get; set; }
     public DateTimeOffset? DeliveredAt { get; set; }
-    public decimal? ReplyToMessageId { get; set; }
+    public long? ReplyToMessageId { get; set; }
 
     public Conversation Conversation { get; set; } = null!;
     public Device SenderDevice { get; set; } = null!;
